+++ ZMath/math/sm4/sm4.h
@@ -2,19 +2,19 @@
   Copyright (C) 2013 by Felipe Tavares
 
   This file is part of Render TRI.
-  
+
   Render TRI is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
   (at your option) any later version.
-  
+
   Render TRI is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
-  
+
   You should have received a copy of the GNU General Public License
-  along with Render TRI.  If not, see <http://www.gnu.org/licenses/>. 
+  along with Render TRI.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #ifndef RI_SM4_H
@@ -33,7 +33,7 @@ public:
   // b.x b.y b.z b.w
   // c.x c.y c.z c.w
   // d.x d.y d.z d.w
- 
+
   sm4 ();
   sm4 (const vec4&,
        const vec4&,
@@ -44,19 +44,19 @@ public:
   sm4 operator + (const sm4&);
   sm4 operator - (const sm4&);
 
-  sm4 operator * (const double);
+  sm4 operator * (const float);
   sm4 operator * (sm4&);
   vec4 operator * (vec4);
 
   sm4 operator += (const sm4&);
   sm4 operator -= (const sm4&);
   sm4 operator *= (sm4&);
-  sm4 operator *= (const double);
+  sm4 operator *= (const float);
 
-  static sm4 rotateX (double);
-  static sm4 rotateY (double);
-  static sm4 rotateZ (double);
-  
+  static sm4 rotateX (float);
+  static sm4 rotateY (float);
+  static sm4 rotateZ (float);
+
 
   sm4 t();
   void identity();
